backend/requirements.txt
Flask
Flask-SQLAlchemy
pandas
plotly
openpyxl
flask-cors
google-generativeai
